% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_post_hoc.R
\name{anova_post_hoc}
\alias{anova_post_hoc}
\title{anova_post_hoc}
\usage{
anova_post_hoc(df)
}
\arguments{
\item{df}{colname is respectively "tr" and "va", which is factor and numeric.}
}
\value{
A list with the anova result value and the post hoc test result.
}
\description{
data anova and the post hoc test
}
\details{
the first col. must be the treatment as factor; the second col. must be the data for analysis.
}
\examples{
library(sxz4stat)
Demo_df <- data.frame(tr=factor(c("A","A","A","A","B","B","B","B","C","C","C","C")),va=c(10,11,11,12,13,14,14,15,16,17,17,18))
anova_post_hoc(Demo_df)
}
