% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demo_df_data.R
\docType{data}
\name{Demo_df}
\alias{Demo_df}
\title{Demo data}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
Demo_df
}
\arguments{
\item{df}{colname is respectively "tr" and "va", which is factor and numeric.}
}
\value{
A list with the anova result value and the post hoc test result.
}
\description{
dataframe
}
\details{
the first col. must be the treatment as factor; the second col. must be the data for analysis.
}
\examples{
Demo_df
}
\keyword{datasets}
