% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_bind.R
\name{df_bind}
\alias{df_bind}
\title{df_bind}
\usage{
df_bind(df = df, gr = gr, ta = ta, col.n = col.n, row.n = row.n)
}
\arguments{
\item{df, }{gr, ta are belonging to dataframe.}
}
\value{
A dataframe was averaged by some group of gr and sumed by some taxonomy of ta.
}
\description{
calculate mean and sum of the df according group of the gr and taxonomy of the ta.
}
\details{
colname of df is fit with the rowname of gr; and rowname of df is fit with the rowname of ta. col.n is one of the colnames of gr, and row.n is one of the colnames of ta.
}
