% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_adjacentMatrix.R
\name{long_to_adjacentMatrix}
\alias{long_to_adjacentMatrix}
\title{long_to_adjacentMatrix}
\usage{
long_to_adjacentMatrix(df = df)
}
\arguments{
\item{df}{is belonging to dataframe.}
}
\value{
A dataframe was formed to easy check.
}
\description{
long data turn to width data; or edge data of network turn to adjacent matrix.
}
\details{
the 1st column of df is the rowname of new dataframe; the 2nd column of df is the colname of new dataframe; the 3rd column of df is the value of new dataframe.
}
